--- components/mypage/mysong/MyVocab.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function MyVocab() {
-  return <div>마이보캡 카드 컴포넌트임</div>;
-}
-
-export default MyVocab;
