+++ components/mypage/myvocab/MyVocab.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function MyVocab() {
+  //   const { data } = useSWR<{ data: IMySongItem[] }>('/user/song', client.get);
+  return <div></div>;
+}
+
+export default MyVocab;
