--- pages/mypage/MyVocab.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function MyVocab() {
-  return <div>MyVocab</div>;
-}
-
-export default MyVocab;
