+++ pages/mypage/collection.tsx
@@ -1,15 +1,12 @@
 import Footer from '@components/common/Footer';
 import Header from '@components/common/Header';
 import Mysong from '@components/mypage/mysong/Mysong';
-import MySongItem from '@components/mypage/mysong/MySongItem';
+import MyVocab from '@components/mypage/myvocab/MyVocab';
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import { mockClient } from 'lib/api';
 import { useRouter } from 'next/router';
 import React, { ReactElement, useEffect, useState } from 'react';
 
-import { IMySongItem } from '../../types';
-import MyVocab from './MyVocab';
 interface StyledProps {
   mySongs: boolean | undefined;
   myVocab: boolean | undefined;
@@ -58,9 +55,7 @@ function Collection(): ReactElement {
           </button>
         </div>
       </Styled.MyCollection>
-      <Styled.DrawCard>
-        <div className="card-item">{pid.type === 'mysongs' ? <Mysong /> : <MyVocab />}</div>
-      </Styled.DrawCard>
+      <Styled.DrawCard>{pid.type === 'mysongs' ? <Mysong /> : <MyVocab />}</Styled.DrawCard>
       <Footer />
     </>
   );
@@ -116,14 +111,5 @@ const Styled = {
     justify-content: center;
     background: #f9fbfd;
     width: 100%;
-
-    .card-item {
-      display: grid;
-      grid-template-columns: repeat(auto-fill, minmax(158px, 1fr));
-      column-gap: 25px;
-      margin: 74px 32px;
-      width: 1070px;
-      row-gap: 25px;
-    }
   `,
 };
